msvc9 compatibility
- sqrt in msvc9 needs explicit cast from int
- freak error with cuda & msvc9 when using modulus operator(%)

+++ pylearn2/sandbox/cuda_convnet/conv_util.cuh
@@ -993,7 +993,7 @@ void convLocalStochasticMaxPool(NVMatrix& images, NVMatrix& target, int numFilte
     int numImages = images.getNumCols();
     int imgPixels = images.getNumRows() / numFilters;
     assert(images.getNumRows() == numFilters * imgPixels);
-    int imgSize = int(sqrt(imgPixels));
+    int imgSize = int(sqrt((double)imgPixels));
     assert(imgSize * imgSize == imgPixels);
     assert(!images.isTrans());
     assert(!target.isTrans());
@@ -1099,7 +1099,7 @@ void convLocalWeightedPool(NVMatrix& images, NVMatrix& target, int numFilters,
     int numImages = images.getNumCols();
     int imgPixels = images.getNumRows() / numFilters;
     assert(images.getNumRows() == numFilters * imgPixels);
-    int imgSize = int(sqrt(imgPixels));
+    int imgSize = int(sqrt((double)imgPixels));
     assert(imgSize * imgSize == imgPixels);
     assert(!images.isTrans());
     assert(!target.isTrans());
